+++ ui.admin-mui/src/routes/commands/customcommandsGroup.tsx
@@ -17,9 +17,12 @@ import { PermissionTypeProvider } from '../../components/Table/PermissionTypePro
 import getAccessToken from '../../getAccessToken';
 import { getPermissionName } from '../../helpers/getPermissionName';
 import { usePermissions } from '../../hooks/usePermissions';
+import { useScope } from '../../hooks/useScope';
 import { useTranslation } from '../../hooks/useTranslation';
 
 const PageCommandsCommands = () => {
+  const scope = useScope('custom_commands');
+
   const { translate } = useTranslation();
   const { id } = useParams();
   const { enqueueSnackbar } = useSnackbar();
@@ -56,44 +59,49 @@ const PageCommandsCommands = () => {
   }, [ groupsSettings, groups ]);
 
   const deleteItem = useCallback((item: CommandsGroup) => {
-    axios.delete(`${JSON.parse(localStorage.server)}/api/systems/customcommands/groups/${item.name}`, { headers: { authorization: `Bearer ${getAccessToken()}` } })
+    axios.delete(`/api/systems/customcommands/groups/${item.name}`, { headers: { authorization: `Bearer ${getAccessToken()}` } })
       .finally(() => {
         enqueueSnackbar(`Commands group ${item.name} deleted successfully. You can still see this group if it is being activelly used by custom commands.`, { variant: 'success' });
         refresh();
       });
   }, [ enqueueSnackbar ]);
 
-  const columns = useMemo<Column[]>(() => [
-    {
-      name:  'name',
-      title: capitalize(translate('group')),
-    },
-    {
-      name:         'used',
-      title:        capitalize(translate('isUsed')),
-      getCellValue: (row) => groups.includes(row.name),
-    },
-    {
-      name:         'filter',
-      title:        capitalize(translate('filter')),
-      getCellValue: (row) => row.options.filter === null ? 'No filter set' : row.options.filter,
-    },
-    {
-      name:         'permission',
-      title:        translate('permission'),
-      getCellValue: (row) => row.options.permission === null ? 'No permission set' : getPermissionName(row.options.permission, permissions || []),
-    },
-    {
-      name:         'actions',
-      title:        ' ',
-      getCellValue: (row) => [
-        <Stack direction="row" key="row">
-          <EditButton href={'/commands/customcommands/group/edit/' + row.name}/>
-          <DeleteButton key='delete' onDelete={() => deleteItem(row)} />
-        </Stack>,
-      ],
-    },
-  ], [ permissions, translate, deleteItem, groups ]);
+  const columns = useMemo<Column[]>(() => {
+    const col: Column[] = [
+      {
+        name:  'name',
+        title: capitalize(translate('group')),
+      },
+      {
+        name:         'used',
+        title:        capitalize(translate('isUsed')),
+        getCellValue: (row) => groups.includes(row.name),
+      },
+      {
+        name:         'filter',
+        title:        capitalize(translate('filter')),
+        getCellValue: (row) => row.options.filter === null ? 'No filter set' : row.options.filter,
+      },
+      {
+        name:         'permission',
+        title:        translate('permission'),
+        getCellValue: (row) => row.options.permission === null ? 'No permission set' : getPermissionName(row.options.permission, permissions || []),
+      }
+    ];
+    if (scope.manage) {
+      col.push({
+        name:         'actions',
+        title:        ' ',
+        getCellValue: (row) => [
+          <Stack direction="row" key="row">
+            <EditButton href={'/commands/customcommands/group/edit/' + row.name}/>
+            <DeleteButton key='delete' onDelete={() => deleteItem(row)} />
+          </Stack>,
+        ],
+      });
+    }
+    return col;
+  }, [ scope, permissions, translate, deleteItem, groups ]);
 
   useEffect(() => {
     refresh().then(() => setLoading(false));
@@ -102,14 +110,14 @@ const PageCommandsCommands = () => {
   const refresh = async () => {
     await Promise.all([
       new Promise<void>(resolve => {
-        axios.get(`${JSON.parse(localStorage.server)}/api/systems/customcommands`, { headers: { authorization: `Bearer ${getAccessToken()}` } })
+        axios.get(`/api/systems/customcommands`, { headers: { authorization: `Bearer ${getAccessToken()}` } })
           .then(({ data }) => {
             setItems(data.data);
             resolve();
           });
       }),
       new Promise<void>(resolve => {
-        axios.get(`${JSON.parse(localStorage.server)}/api/systems/customcommands/groups`, { headers: { authorization: `Bearer ${getAccessToken()}` } })
+        axios.get(`/api/systems/groups/customcommands`, { headers: { authorization: `Bearer ${getAccessToken()}` } })
           .then(({ data }) => {
             setGroupsSettings(data.data);
             resolve();
